v0.1 of eventviewer done
