+++ .dockerignore
@@ -1,2 +1,1 @@
-/Data
 */__pycache__
