+++ Config/supervisord.conf
@@ -7,7 +7,7 @@ port=*:9001        ; (ip_address:port specifier, *:port for all iface)
 ;password=daedalus               ; (default is no password (open server))
 
 [program:evk_horizon]
-command= python3 record_raw_ekv4_w_temp_and_illum.py --recordings /root/data/evk4_horizon --route horizon --port 8000 %(ENV_SERIAL_1)s
+command= python3 record_raw_ekv4_w_temp_and_illum.py --recordings /root/data/evk4_horizon %(ENV_SERIAL_1)s
 directory=/root/code
 autorestart=true
 startretries=10000
@@ -25,7 +25,7 @@ autorestart=true
 startretries=10000
 
 [program:evk_space]
-command= python3 record_raw_ekv4_w_temp_and_illum.py --recordings /root/data/evk4_space --route space --port 8001 %(ENV_SERIAL_0)s
+command= python3 record_raw_ekv4_w_temp_and_illum.py --recordings /root/data/evk4_space %(ENV_SERIAL_0)s
 directory=/root/code
 autorestart=true
 startretries=10000
@@ -48,3 +48,8 @@ directory=/root/code
 autorestart=true
 startretries=10000
 stdout_logfile=/var/log/supervisor/%(program_name)s.log
+
+[program:camera_tester]
+command=python3 eventviewer_server.py --route horizon --port 8000 %(ENV_SERIAL_0)s
+directory=/root/code
+autostart=false
